+++ src/CRC.h
@@ -22,8 +22,8 @@ uint8_t calcCRC8(
     uint8_t polynome   = CRC8_POLYNOME,
     uint8_t initial    = CRC8_INITIAL,
     uint8_t xorOut     = CRC8_XOR_OUT,
-    bool reverseIn     = CRC8_REF_IN,
-    bool reverseOut    = CRC8_REF_OUT,
+    bool reverseIn     = CRC8_REV_IN,
+    bool reverseOut    = CRC8_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 uint16_t calcCRC12(
@@ -31,8 +31,8 @@ uint16_t calcCRC12(
     uint16_t polynome  = CRC12_POLYNOME,
     uint16_t initial   = CRC12_INITIAL,
     uint16_t xorOut    = CRC12_XOR_OUT,
-    bool reverseIn     = CRC12_REF_IN,
-    bool reverseOut    = CRC12_REF_OUT,
+    bool reverseIn     = CRC12_REV_IN,
+    bool reverseOut    = CRC12_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 uint16_t calcCRC16(
@@ -40,8 +40,8 @@ uint16_t calcCRC16(
     uint16_t polynome  = CRC16_POLYNOME,
     uint16_t initial   = CRC16_INITIAL,
     uint16_t xorOut    = CRC16_XOR_OUT,
-    bool reverseIn     = CRC16_REF_IN,
-    bool reverseOut    = CRC16_REF_OUT,
+    bool reverseIn     = CRC16_REV_IN,
+    bool reverseOut    = CRC16_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 uint32_t calcCRC32(
@@ -49,8 +49,8 @@ uint32_t calcCRC32(
     uint32_t polynome  = CRC32_POLYNOME,
     uint32_t initial   = CRC32_INITIAL,
     uint32_t xorOut    = CRC32_XOR_OUT,
-    bool reverseIn     = CRC32_REF_IN,
-    bool reverseOut    = CRC32_REF_OUT,
+    bool reverseIn     = CRC32_REV_IN,
+    bool reverseOut    = CRC32_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 uint64_t calcCRC64(
@@ -58,8 +58,8 @@ uint64_t calcCRC64(
     uint64_t polynome  = CRC64_POLYNOME,
     uint64_t initial   = CRC64_INITIAL,
     uint64_t xorOut    = CRC64_XOR_OUT,
-    bool reverseIn     = CRC64_REF_IN,
-    bool reverseOut    = CRC64_REF_OUT,
+    bool reverseIn     = CRC64_REV_IN,
+    bool reverseOut    = CRC64_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 
@@ -69,8 +69,8 @@ uint8_t crc8(
     uint8_t polynome   = CRC8_POLYNOME,
     uint8_t initial    = CRC8_INITIAL,
     uint8_t xorOut     = CRC8_XOR_OUT,
-    bool reverseIn     = CRC8_REF_IN,
-    bool reverseOut    = CRC8_REF_OUT,
+    bool reverseIn     = CRC8_REV_IN,
+    bool reverseOut    = CRC8_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 [[deprecated("Use calcCRC12() instead")]]
@@ -79,8 +79,8 @@ uint16_t crc12(
     uint16_t polynome  = CRC12_POLYNOME,
     uint16_t initial   = CRC12_INITIAL,
     uint16_t xorOut    = CRC12_XOR_OUT,
-    bool reverseIn     = CRC12_REF_IN,
-    bool reverseOut    = CRC12_REF_OUT,
+    bool reverseIn     = CRC12_REV_IN,
+    bool reverseOut    = CRC12_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 [[deprecated("Use calcCRC16() instead")]]
@@ -89,8 +89,8 @@ uint16_t crc16(
     uint16_t polynome  = CRC16_POLYNOME,
     uint16_t initial   = CRC16_INITIAL,
     uint16_t xorOut    = CRC16_XOR_OUT,
-    bool reverseIn     = CRC16_REF_IN,
-    bool reverseOut    = CRC16_REF_OUT,
+    bool reverseIn     = CRC16_REV_IN,
+    bool reverseOut    = CRC16_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 [[deprecated("Use calcCRC16() with CRC16_CCITT_FALSE params instead")]]
@@ -99,8 +99,8 @@ uint16_t crc16_CCITT(
     uint16_t polynome  = CRC16_CCITT_FALSE_POLYNOME,
     uint16_t initial   = CRC16_CCITT_FALSE_INITIAL,
     uint16_t xorOut    = CRC16_CCITT_FALSE_XOR_OUT,
-    bool reverseIn     = CRC16_CCITT_FALSE_REF_IN,
-    bool reverseOut    = CRC16_CCITT_FALSE_REF_OUT,
+    bool reverseIn     = CRC16_CCITT_FALSE_REV_IN,
+    bool reverseOut    = CRC16_CCITT_FALSE_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 [[deprecated("Use calcCRC32() instead")]]
@@ -109,8 +109,8 @@ uint32_t crc32(
     uint32_t polynome  = CRC32_POLYNOME,
     uint32_t initial   = CRC32_INITIAL,
     uint32_t xorOut    = CRC32_XOR_OUT,
-    bool reverseIn     = CRC32_REF_IN,
-    bool reverseOut    = CRC32_REF_OUT,
+    bool reverseIn     = CRC32_REV_IN,
+    bool reverseOut    = CRC32_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);
 
 [[deprecated("Use calcCRC64() instead")]]
@@ -119,8 +119,8 @@ uint64_t crc64(
     uint64_t polynome  = CRC64_POLYNOME,
     uint64_t initial   = CRC64_INITIAL,
     uint64_t xorOut    = CRC64_XOR_OUT,
-    bool reverseIn     = CRC64_REF_IN,
-    bool reverseOut    = CRC64_REF_OUT,
+    bool reverseIn     = CRC64_REV_IN,
+    bool reverseOut    = CRC64_REV_OUT,
     crc_size_t yieldPeriod = CRC_YIELD_DISABLED);