+++ src/CRC12.h
@@ -16,14 +16,14 @@ public:
   CRC12(uint16_t polynome = CRC12_POLYNOME,
         uint16_t initial  = CRC12_INITIAL,
         uint16_t xorOut   = CRC12_XOR_OUT,
-        bool reverseIn    = CRC12_REF_IN,
-        bool reverseOut   = CRC12_REF_OUT);
+        bool reverseIn    = CRC12_REV_IN,
+        bool reverseOut   = CRC12_REV_OUT);
 
   void reset(uint16_t polynome = CRC12_POLYNOME,
              uint16_t initial  = CRC12_INITIAL,
              uint16_t xorOut   = CRC12_XOR_OUT,
-             bool reverseIn    = CRC12_REF_IN,
-             bool reverseOut   = CRC12_REF_OUT);
+             bool reverseIn    = CRC12_REV_IN,
+             bool reverseOut   = CRC12_REV_OUT);
 
   void restart();
   uint16_t calc() const;