+++ src/CRC16.h
@@ -15,14 +15,14 @@ public:
   CRC16(uint16_t polynome = CRC16_POLYNOME,
         uint16_t initial  = CRC16_INITIAL,
         uint16_t xorOut   = CRC16_XOR_OUT,
-        bool reverseIn    = CRC16_REF_IN,
-        bool reverseOut   = CRC16_REF_OUT);
+        bool reverseIn    = CRC16_REV_IN,
+        bool reverseOut   = CRC16_REV_OUT);
 
   void reset(uint16_t polynome = CRC16_POLYNOME,
              uint16_t initial  = CRC16_INITIAL,
              uint16_t xorOut   = CRC16_XOR_OUT,
-             bool reverseIn    = CRC16_REF_IN,
-             bool reverseOut   = CRC16_REF_OUT);
+             bool reverseIn    = CRC16_REV_IN,
+             bool reverseOut   = CRC16_REV_OUT);
 
   void restart();
   uint16_t calc() const;