+++ src/CRC32.h
@@ -15,14 +15,14 @@ public:
   CRC32(uint32_t polynome = CRC32_POLYNOME,
         uint32_t initial  = CRC32_INITIAL,
         uint32_t xorOut   = CRC32_XOR_OUT,
-        bool reverseIn    = CRC32_REF_IN,
-        bool reverseOut   = CRC32_REF_OUT);
+        bool reverseIn    = CRC32_REV_IN,
+        bool reverseOut   = CRC32_REV_OUT);
 
   void reset(uint32_t polynome = CRC32_POLYNOME,
              uint32_t initial  = CRC32_INITIAL,
              uint32_t xorOut   = CRC32_XOR_OUT,
-             bool reverseIn    = CRC32_REF_IN,
-             bool reverseOut   = CRC32_REF_OUT);
+             bool reverseIn    = CRC32_REV_IN,
+             bool reverseOut   = CRC32_REV_OUT);
 
   void restart();
   uint32_t calc() const;