+++ src/CRC8.h
@@ -14,14 +14,14 @@ public:
   CRC8(uint8_t polynome = CRC8_POLYNOME,
        uint8_t initial  = CRC8_INITIAL,
        uint8_t xorOut   = CRC8_XOR_OUT,
-       bool reverseIn   = CRC8_REF_IN,
-       bool reverseOut  = CRC8_REF_OUT);
+       bool reverseIn   = CRC8_REV_IN,
+       bool reverseOut  = CRC8_REV_OUT);
 
   void reset(uint8_t polynome = CRC8_POLYNOME,
              uint8_t initial  = CRC8_INITIAL,
              uint8_t xorOut   = CRC8_XOR_OUT,
-             bool reverseIn   = CRC8_REF_IN,
-             bool reverseOut  = CRC8_REF_OUT);
+             bool reverseIn   = CRC8_REV_IN,
+             bool reverseOut  = CRC8_REV_OUT);
 
   void restart();
   uint8_t calc() const;